requirements.txt
vosk
pygame
coqui-tts
gpiozero